+++ zu-plus.php
@@ -4,7 +4,7 @@ Plugin Name: ZU+
 Plugin URI: https://dmitryrudakov.ru/plugins/
 GitHub Plugin URI: https://github.com/picasso/zu-plus
 Description: This plugin encompasses ZU framework functionality.
-Version: 1.1.9
+Version: 1.2.0
 Author: Dmitry Rudakov
 Author URI: https://dmitryrudakov.ru/about/
 Text Domain: zu-plugin
@@ -32,7 +32,7 @@ Domain Path: /lang/
 
 // Prohibit direct script loading
 defined('ABSPATH') || die('No direct script access allowed!');
-define('ZUPLUS_VERSION', '1.1.9');
+define('ZUPLUS_VERSION', '1.2.0');
 define('ZUPLUS_NAME', 'ZU+');
 define('__ZUPLUS_ROOT__', plugin_dir_path(__FILE__)); 
 define('__ZUPLUS_FILE__', __FILE__); 
@@ -269,7 +269,7 @@ class ZU_Admin extends zuplus_Admin {
 		echo $this->form->fields('Duplicate Page Settings.');
 	}
 	
-	public function ajax_more($option_name) {
+	public function ajax_more($option_name, $ajax_value) {
 		if($option_name === 'zuplus_clear_log') return $this->plugin->dbug->clear_log();
 		if($option_name === 'zuplus_duplicate_menu') return zuplus_ajax_duplicate_menu();